@@ -3,85 +3,93 @@
   "version": "0.0.1",
   "license": "MIT",
   "scripts": {
-    "start": "nx serve",
+    "start": "node server.js",
     "build": "nx build",
-    "test": "nx test"
+    "test": "nx test",
+    "heroku-postbuild": "nx build"
+  },
+  "engines": {
+    "node": "20.10.0",
+    "npm": "10.2.4"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "17.1.1",
-    "@angular/cdk": "17.1.1",
-    "@angular/common": "17.1.1",
-    "@angular/compiler": "17.1.1",
-    "@angular/core": "17.1.1",
-    "@angular/forms": "17.1.1",
-    "@angular/material": "17.1.1",
-    "@angular/platform-browser": "17.1.1",
-    "@angular/platform-browser-dynamic": "17.1.1",
-    "@angular/router": "17.1.1",
+    "@angular/animations": "18.0.4",
+    "@angular/cdk": "18.0.4",
+    "@angular/common": "18.0.4",
+    "@angular/compiler": "18.0.4",
+    "@angular/core": "18.0.4",
+    "@angular/forms": "18.0.4",
+    "@angular/material": "18.0.4",
+    "@angular/platform-browser": "18.0.4",
+    "@angular/platform-browser-dynamic": "18.0.4",
+    "@angular/router": "18.0.4",
     "@aws-amplify/ui-angular": "^5.0.8",
     "@ngx-translate/core": "^15.0.0",
     "@ngx-translate/http-loader": "8.0.0",
-    "@ngxs/store": "3.8.2",
-    "@nx/angular": "17.3.0",
-    "@rx-angular/cdk": "17.0.0",
-    "@rx-angular/template": "17.0.0",
-    "ably": "^1.2.48",
-    "angular-feather": "^6.5.1",
-    "aws-amplify": "^6.0.13",
-    "rxjs": "~7.8.1",
-    "tailwindcss-animated": "^1.0.1",
+    "@ngxs/store": "18.0.0",
+    "@nx/angular": "19.3.1",
+    "@rx-angular/cdk": "18.0.0",
+    "@rx-angular/template": "18.0.0",
+    "ably": "2.1.0",
+    "angular-feather": "6.5.1",
+    "aws-amplify": "6.3.7",
+    "express": "^4.18.3",
+    "path": "0.12.7",
+    "rxjs": "7.8.1",
+    "tailwindcss-animated": "1.1.2",
     "tslib": "^2.6.2",
     "zone.js": "0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "17.1.1",
-    "@angular-devkit/core": "17.1.1",
-    "@angular-devkit/schematics": "17.1.1",
-    "@angular-eslint/eslint-plugin": "17.1.1",
-    "@angular-eslint/eslint-plugin-template": "17.1.1",
-    "@angular-eslint/template-parser": "17.1.1",
-    "@angular/cli": "~17.1.1",
-    "@angular/compiler-cli": "17.1.1",
-    "@angular/language-service": "17.1.1",
-    "@nx/cypress": "17.3.0",
-    "@nx/eslint-plugin": "17.3.0",
-    "@nx/jest": "17.3.0",
-    "@nx/js": "17.3.0",
-    "@nx/storybook": "17.3.0",
-    "@nx/web": "17.3.0",
-    "@nx/workspace": "17.3.0",
-    "@schematics/angular": "17.1.1",
-    "@storybook/addon-essentials": "7.5.1",
-    "@storybook/angular": "7.6.0",
-    "@storybook/core-server": "7.6.0",
+    "@angular-devkit/build-angular": "18.0.5",
+    "@angular-devkit/core": "18.0.5",
+    "@angular-devkit/schematics": "18.0.5",
+    "@angular-eslint/eslint-plugin": "18.0.1",
+    "@angular-eslint/eslint-plugin-template": "18.0.1",
+    "@angular-eslint/template-parser": "18.0.1",
+    "@angular/cli": "~18.0.0",
+    "@angular/compiler-cli": "18.0.4",
+    "@angular/language-service": "18.0.4",
+    "@nx/cypress": "19.3.1",
+    "@nx/eslint": "19.3.1",
+    "@nx/eslint-plugin": "19.3.1",
+    "@nx/jest": "19.3.1",
+    "@nx/js": "19.3.1",
+    "@nx/storybook": "19.3.1",
+    "@nx/web": "19.3.1",
+    "@nx/workspace": "19.3.1",
+    "@schematics/angular": "18.0.5",
+    "@storybook/addon-essentials": "8.1.10",
+    "@storybook/angular": "8.1.10",
+    "@storybook/core-server": "8.1.10",
     "@types/jest": "^29.5.11",
     "@types/node": "18.19.10",
-    "@typescript-eslint/eslint-plugin": "6.20.0",
-    "@typescript-eslint/parser": "6.20.0",
+    "@typescript-eslint/eslint-plugin": "7.9.0",
+    "@typescript-eslint/parser": "7.9.0",
+    "@typescript-eslint/utils": "^8.0.0-alpha.28",
     "autoprefixer": "^10.4.17",
-    "cypress": "^13.6.3",
-    "eslint": "~8.56.0",
+    "cypress": "13.9.0",
+    "eslint": "8.57.0",
     "eslint-config-prettier": "9.1.0",
-    "eslint-plugin-cypress": "^2.15.1",
-    "eslint-plugin-prefer-arrow-functions": "^3.2.4",
+    "eslint-plugin-cypress": "3.3.0",
+    "eslint-plugin-prefer-arrow-functions": "3.3.2",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
-    "jest-preset-angular": "14.0.0",
+    "jest-preset-angular": "14.1.1",
     "jsonc-eslint-parser": "^2.4.0",
-    "ng-packagr": "17.1.2",
-    "nx": "17.3.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "ng-packagr": "18.0.0",
+    "nx": "19.3.1",
     "postcss": "^8.4.33",
     "postcss-import": "~14.1.0",
     "postcss-preset-env": "~7.5.0",
     "postcss-url": "~10.1.3",
     "prettier": "^2.6.2",
-    "tailwindcss": "3.4.1",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
+    "tailwindcss": "3.4.4",
     "ts-jest": "^29.1.2",
     "ts-node": "10.9.2",
-    "typescript": "5.3.3",
-    "@nx/eslint": "17.3.0"
+    "typescript": "5.4.5"
   }
 }
