@@ -1,10 +1,10 @@
 import { DATE_PIPE_DEFAULT_OPTIONS, registerLocaleData } from '@angular/common';
-import { HTTP_INTERCEPTORS, HttpClient, HttpClientModule } from '@angular/common/http';
+import { HTTP_INTERCEPTORS, HttpClient, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import localeDe from '@angular/common/locales/de';
 import localeDeExtra from '@angular/common/locales/extra/de';
 import localeItExtra from '@angular/common/locales/extra/it';
 import localeIt from '@angular/common/locales/it';
-import { LOCALE_ID, NgModule } from '@angular/core';
+import { APP_INITIALIZER, LOCALE_ID, NgModule } from '@angular/core';
 import { MAT_RIPPLE_GLOBAL_OPTIONS, MatNativeDateModule, RippleGlobalOptions } from '@angular/material/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
@@ -12,7 +12,9 @@ import { AmplifyAuthenticatorModule } from '@aws-amplify/ui-angular';
 import { TranslateLoader, TranslateModule, TranslateService } from '@ngx-translate/core';
 import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 import { NgxsModule } from '@ngxs/store';
-import { ActionNotificationModule, AUTH_JWT_TOKEN_FACTORY, BASE_API_URL, ERROR_ACTION_EXECUTOR, SnackbarService } from '@vyf/base';
+import { ABLY_TOKEN_URL_FACTORY, ActionNotificationModule, AUTH_JWT_TOKEN_FACTORY, ERROR_ACTION_EXECUTOR, SnackbarService } from '@vyf/base';
+import { USER_API_URL } from '@vyf/user-service';
+import { VOTE_CIRCLE_API_URL } from '@vyf/vote-circle-service';
 import { Amplify } from 'aws-amplify';
 import awsconfig from '../aws-exports';
 import { environment } from '../env/environment';
@@ -21,15 +23,16 @@ import { AppComponent } from './app.component';
 import { AuthInterceptor } from './core/auth/auth-interceptor.interceptor';
 import { actionErroredFactory } from './core/factory/action-errored.factory';
 import { authJwtTokenFactory } from './core/factory/auth-jw-token.factory';
+import { ablyTokenEndpointUrlFactory, userBaseApiUrlFactory, voteCircleBaseApiUrlFactory } from './core/factory/base-url.factory';
+import { initializeAppFactory } from './core/factory/init-app.factory';
 import { langCodeFactory, timezoneFactory } from './core/factory/lang-code.factory';
 import { AwsCognitoService } from './core/services/aws-cognito.service';
+import { UserStorageService } from './core/services/user-storage.service';
 import { CirclesErrorTrackedActions } from './modules/circles/state/actions/circles.action';
-import { LayoutModule } from './modules/layout/layout.module';
 import { RankingErrorTrackedActions } from './modules/ranking/state/actions/ranking.action';
-import { UserErrorTrackedActions } from './modules/user/state/actions/user.action';
-import { UserState } from './modules/user/state/user.state';
 import { InfoErrorTrackedActions } from './shared/state/actions/info.action';
 import { MemberCircleErrorTrackedActions, MemberErrorTrackedActions, MemberRankingErrorTrackedActions } from './shared/state/actions/member.action';
+import { UserErrorTrackedActions } from './shared/state/actions/user.action';
 
 registerLocaleData(localeDe, 'de', localeDeExtra);
 registerLocaleData(localeIt, 'it', localeItExtra);
@@ -46,12 +49,9 @@ const globalRippleConfig: RippleGlobalOptions = {
 
 @NgModule({
     declarations: [AppComponent],
-    imports: [
-        BrowserModule,
+    bootstrap: [AppComponent], imports: [BrowserModule,
         BrowserAnimationsModule,
-        HttpClientModule,
-        AmplifyAuthenticatorModule,
-        NgxsModule.forRoot([UserState], {
+        NgxsModule.forRoot([], {
             developmentMode: !environment.production
         }),
         ActionNotificationModule.forRoot({
@@ -65,21 +65,18 @@ const globalRippleConfig: RippleGlobalOptions = {
                 ...InfoErrorTrackedActions
             ]
         }),
-        TranslateModule.forRoot(
-            {
-                loader: {
-                    provide: TranslateLoader,
-                    useFactory: (http: HttpClient) => new TranslateHttpLoader(http, './assets/i18n/core/'),
-                    deps: [HttpClient]
-                },
-                extend: true
-            }
-        ),
-        LayoutModule,
+        TranslateModule.forRoot({
+            loader: {
+                provide: TranslateLoader,
+                useFactory: (http: HttpClient) => new TranslateHttpLoader(http, './assets/i18n/core/'),
+                deps: [HttpClient]
+            },
+            extend: true
+        }),
+        AmplifyAuthenticatorModule,
         AppRoutingModule,
-        MatNativeDateModule
-    ],
-    providers: [
+        MatNativeDateModule], providers: [
+        UserStorageService,
         AwsCognitoService,
         {
             provide: HTTP_INTERCEPTORS,
@@ -87,8 +84,26 @@ const globalRippleConfig: RippleGlobalOptions = {
             multi: true
         },
         {
-            provide: BASE_API_URL,
-            useValue: ''
+            provide: APP_INITIALIZER,
+            useFactory: initializeAppFactory,
+            multi: true,
+            deps: [AwsCognitoService, UserStorageService]
+        },
+        // {
+        //     provide: BASE_API_URL,
+        //     useValue: ''
+        // },
+        {
+            provide: USER_API_URL,
+            useFactory: userBaseApiUrlFactory
+        },
+        {
+            provide: VOTE_CIRCLE_API_URL,
+            useFactory: voteCircleBaseApiUrlFactory
+        },
+        {
+            provide: ABLY_TOKEN_URL_FACTORY,
+            useFactory: ablyTokenEndpointUrlFactory
         },
         {
             provide: MAT_RIPPLE_GLOBAL_OPTIONS,
@@ -112,8 +127,8 @@ const globalRippleConfig: RippleGlobalOptions = {
         {
             provide: DATE_PIPE_DEFAULT_OPTIONS,
             useFactory: timezoneFactory
-        }
-    ],
-    bootstrap: [AppComponent]
+        },
+        provideHttpClient(withInterceptorsFromDi())
+    ]
 })
 export class AppModule {}
