@@ -2,6 +2,5 @@ export const Route = {
     default: '',
     user: 'user',
     circles: 'circles',
-    ranking: 'ranking',
-    members: 'members'
+    ranking: 'ranking'
 };
